shorts-generator/requirements.txt
gradio>=5.0.0
modal>=0.53.0
pytube>=15.0.0
ffmpeg-python>=0.2.0
openai>=1.0.0
openai-whisper>=20231117
python-dotenv>=1.0.0
langchain>=0.1.0
google-generativeai>=0.3.0
langchain-google-genai>=0.0.5
chromadb>=0.4.18
sentence-transformers>=2.2.2
yt-dlp>=2023.3.4
numpy>=1.20.0
requests>=2.28.0